requirements.txt
asgiref==3.8.1
boto3==1.37.27
botocore==1.37.27
crispy-bootstrap5==2024.10
crispy-tailwind==1.0.3
Django==5.1.7
django-allauth==65.5.0
django-crispy-forms==2.3
django-extensions==3.2.3
django-storages==1.14.2
docutils==0.21.2
jmespath==1.0.1
pillow==11.1.0
psycopg2-binary==2.9.10
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-http-client==3.3.7
s3transfer==0.11.4
sendgrid==6.11.0
sendgrid-django==4.2.0
six==1.17.0
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
urllib3==2.3.0
